+++ src/Context/CartContext/index.js
@@ -1,0 +1,1 @@
+export { CartContextProvider, useCartContext } from "./CartContext";
