+++ src/Context/SnakbarContext/SnakbarContext.jsx
@@ -1,30 +1,11 @@
 import { createContext, useReducer, useContext } from "react";
-
+import { reducer } from "./reducer";
 const SnakbarContext = createContext();
 
-const snakbarReducer = (state, action) => {
-  switch (action.type) {
-    case "INITAIL":
-      return { isShow: false, alertType: "" };
-    case "DEFAULT":
-      return { isShow: true, alertType: "DEFAULT", msg: action.payload };
-    case "ERROR":
-      return { isShow: true, alertType: "ERROR", msg: action.payload };
-    case "WARNING":
-      return { isShow: true, alertType: "WARNING", msg: action.payload };
-    case "SUCCESS":
-      return { isShow: true, alertType: "SUCCESS", msg: action.payload };
-    default:
-      return state;
-  }
-};
 const intialSnakbar = { isShow: false, alertType: "", msg: "" };
 
 export const SnakbarContextProvider = ({ children }) => {
-  const [snakbarStatus, snakbarDispatch] = useReducer(
-    snakbarReducer,
-    intialSnakbar
-  );
+  const [snakbarStatus, snakbarDispatch] = useReducer(reducer, intialSnakbar);
 
   return (
     <SnakbarContext.Provider
