--- src/Context/index.js
@@ -1,10 +1,0 @@
-export { CartContextProvider, useCartContext } from "./CartContext/CartContext";
-export {
-  LoaderContextProvider,
-  useStatus,
-} from "./LoaderContext/LoaderContext";
-export {
-  SnakbarContextProvider,
-  useSnakbarContext,
-} from "./SnakbarContext/SnakbarContext";
-export { AuthProvider, useAuthContext } from "./AuthContext";
