+++ src/components/Snakbar/Snakbar.jsx
@@ -1,7 +1,7 @@
+import "./snakbar.css";
 import { useEffect } from "react";
 import { CloseIcon, CheckIcon } from "../../assests/icons";
-import { useSnakbarContext } from "../../Context";
-import "./snakbar.css";
+import { useSnakbarContext } from "../../Context/SnakbarContext";
 export const Snakbar = () => {
   const { snakbarStatus, snakbarDispatch } = useSnakbarContext();
   useEffect(() => {
