+++ src/pages/Profile/Profile.jsx
@@ -1,10 +1,10 @@
-import { useAuthContext } from "../../Context";
+import { useAuthContext } from "../../Context/AuthContext";
 export const Profile = () => {
-  const { userLogout } = useAuthContext();
+  const { handleLogout } = useAuthContext();
   return (
     <>
       <section className="row justify-center align-center padding-64">
-        <button className="sm-btn-pry-fil btn-err" onClick={userLogout}>
+        <button className="sm-btn-pry-fil btn-err" onClick={handleLogout}>
           Logout
         </button>
       </section>
